soft-launch, add favicons, some share og
