+++ .env.example
@@ -1,1 +1,5 @@
+export NEXT_PUBLIC_ORIGIN=http://localhost:8080
+export NEXT_PUBLIC_SITE_TITLE='Level Protocol'
+export NEXT_PUBLIC_SITE_DESCRIPTION='lvl is a crypto resume: Level is an on-chain reputation and skills web3 resume that highlights all of your contributions across communities, DAOs, and metaverses'
+export NEXT_PUBLIC_TWITTER_USERNAME='lvlprotocol'
 export CHROMATIC_PROJECT_TOKEN=token
