+++ public/images/browserconfig.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<browserconfig>
+  <msapplication>
+    <tile>
+      <square70x70logo src="/images/favicons/tile-70x70.png" />
+      <square150x150logo src="/images/favicons/tile-150x150.png" />
+      <square310x310logo src="/images/favicons/tile-310x310.png" />
+      <wide310x150logo src="/images/favicons/tile-310x150.png" />
+      <TileColor>#d7d7d7</TileColor>
+    </tile>
+  </msapplication>
+</browserconfig>
