+++ util/theme.js
@@ -46,10 +46,6 @@ const theme = {
     mutedBlue: '#6c7894',
     black: '#000000',
     white: '#ffffff',
-    base900: '#0d1034', // deprecated
-    base700: '#3b4472', // deprecated
-    base100: '#e1e4f0', // deprecated
-    primary100: '#cafee8', // deprecated
   },
 
   halftones: {
